Updated serialisation rules with additional setting to control it
JsonNameAttribute now is a child class of JsonMemberAttribute class,
thus JsonOptIn attribute will utilises fields and properties with JsonName
attribute as well. This adds some flexibilities to manipulate the json file.

For controlling the private variables in json data class, A new setting
SearchPrivate is introduced to JsonWriterSettings. Non public fields
and properties will be automatically serialised when SearchPrivate is
enabled. Also, non public fields and properties now will be serialised if
given JsonMember or JsonName attribute regardless of SearchPrivate
setting.

Some comments are changed to match the new updates.